[refactor] : 신고 승인 및 조회 구현체 분리
